entrypoint: rewrite the function call preparation
The previous code used `exec()` with a hacked together string to call
the right function, and that could fail in some circumstances (eg.
Windows with the rs/rsne functions) and lead to security issues. With
this change we now build a proper argument and keyword arguments list
and call the right function by reference. The logic has also been
slightly refactored.